Add files via upload
In this Pricing Page project, I have used  HTML , CSSand JavaScript  to create a visually appealing and user-friendly pricing page for a fictional product or service. 
The project includes building the page's structure and style using HTML and CSS, ensuring a responsive design that adapts to different screen sizes and  incorporating JavaScript for dynamic elements such as interactive pricing toggles or hover effects.